Changed CardFilter to TopInfoBar and created new CardFilter component
